requirements.txt
opencv-python>=4.8.0
mss>=9.0.0
Pillow>=10.0.0
numpy>=1.24.0
google-generativeai>=0.3.0
